FISH-6976: adding validation for mp_scope and mp_app
