+++ appserver/payara-appserver-modules/microprofile/metrics/src/main/java/fish/payara/microprofile/metrics/cdi/MetricUtils.java
@@ -127,17 +127,6 @@ public final class MetricUtils<T extends Metric> {
         return (T) getOrRegister(metric).byMetadataAndTags.apply(registry, metadata, tags);
     }
 
-    public static MetricID validateAndComplementTags(MetricID metricID, String scope) {
-        Map<String, String> tags =metricID.getTags();
-        Optional<String> optionalKey = tags.keySet().stream().filter(k -> k.equals("mp_scope")).findAny();
-        if(!optionalKey.isPresent()) {
-            Tag t = new Tag("mp_scope", scope);
-            MetricID newMetricID = new MetricID(metricID.getName(), t);
-            return newMetricID;
-        }
-        return metricID;
-    }
-
     public static Tag[] setScopeTagForMetric(String scope, Tag... tags) {
         Tag[] tArray = new Tag[1];
         if(scope.equals(MetricRegistry.BASE_SCOPE)) {
@@ -181,7 +170,7 @@ public final class MetricUtils<T extends Metric> {
 
     @SuppressWarnings("unchecked")
     private static Gauge<?> getGauge(MetricRegistry registry, String name, Tag[] tags) {
-        MetricID complementedMetricID = new MetricID(name, tags);//MetricUtils.validateAndComplementTags();
+        MetricID complementedMetricID = new MetricID(name, tags);
         Gauge<?> gauge = registry.getGauges().get(complementedMetricID);
         return gauge != null ? gauge : new LazyGauge<>(() -> registry.getGauges().get(complementedMetricID));
     }
